requirements.txt
pyrogram==2.0.106
python-dotenv==1.0.0
openai==2.24.0
requests==2.32.5
flask==2.3.3